requirements.txt
# Web & API
fastapi==0.110.1
uvicorn==0.29.0
python-multipart==0.0.9
websockets>=12.0

# Groq (STT - free Whisper)
groq>=0.5.0

# ElevenLabs (TTS)
elevenlabs>=1.0.0

# Google Gemini (LLM + embeddings)
google-generativeai>=0.5.2

# Pinecone Vector DB
pinecone-client>=4.0.0

# PDF & Vision
PyMuPDF>=1.24.1
Pillow>=10.3.0

# LangChain
langchain>=0.1.16
langchain-google-genai>=1.0.3
langchain-pinecone>=0.1.0
langchain-community>=0.0.34
langchain-text-splitters>=0.0.1
numpy<2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
boto3>=1.34.0

# Local ASR Dependencies
faster-whisper==1.0.3
torch==2.3.0
torchaudio==2.3.0

